+++ lib/main.js
@@ -1,7 +1,7 @@
 'use babel'
 
 import {Emitter, CompositeDisposable} from 'atom'
-import include from './mixins/include'
+import include from './decorators/include'
 import Minimap from './minimap'
 import MinimapElement from './minimap-element'
 import PluginManagement from './mixins/plugin-management'
