+++ lib/minimap-element.js
@@ -2,7 +2,7 @@
 
 import {CompositeDisposable, Disposable} from 'atom'
 import {registerOrUpdateElement, EventsDelegation, AncestorsMethods} from 'atom-utils'
-import include from './mixins/include'
+import include from './decorators/include'
 import DOMStylesReader from './mixins/dom-styles-reader'
 import CanvasDrawer from './mixins/canvas-drawer'
 import MinimapQuickSettingsElement from './minimap-quick-settings-element'
