+++ lib/minimap.js
@@ -1,7 +1,7 @@
 'use babel'
 
 import {Emitter, CompositeDisposable} from 'atom'
-import include from './mixins/include'
+import include from './decorators/include'
 import DecorationManagement from './mixins/decoration-management'
 import LegacyAdater from './adapters/legacy-adapter'
 import StableAdapter from './adapters/stable-adapter'
